requirements.txt
accelerate==1.8.1
langchain==0.3.26
pillow==11.3.0
qwen-vl-utils[decord]==0.0.8
torch==2.7.1
torchvision==0.22.1
transformers==4.53.2
streamlit==1.46.1